requirements/base.txt

fastapi==0.99.1
httpx[http2]==0.24.1
python-dotenv==0.21.0
uvicorn[standard]==0.21.1